Increase to 3.1, Fix Issue #28

+++ PgBulkInsert/src/main/java/de/bytefish/pgbulkinsert/pgsql/handlers/ByteValueHandler.java
@@ -10,6 +10,6 @@ public class ByteValueHandler<T extends Number> extends BaseValueHandler<T> {
     @Override
     protected void internalHandle(DataOutputStream buffer, final T value) throws Exception {
         buffer.writeInt(1);
-        buffer.writeInt(value.byteValue());
+        buffer.writeByte(value.byteValue());
     }
 }
